+++ src/datamodel/task.h
@@ -14,7 +14,7 @@ enum repetition {Taeglich, Woechentlich, Monatlich, Jaehrlich, keine};
 
 class task{
     private:
-        int identifier;
+        unsigned long identifier;
         string title;
         string description;
         priority tpriority;
@@ -26,9 +26,9 @@ class task{
 
     public:
         task();
-        task(int identifier, string title, string description, priority tpriority, tdate date, ttime startingtime, int length, repetition trepetition, string category);     //if not set, empty string?
+        task(unsigned long identifier, string title, string description, priority tpriority, tdate date, ttime startingtime, int length, repetition trepetition, string category);     //if not set, empty string?
 
-        void set_identifier(int identifier);
+        void set_identifier(unsigned long identifier);
         void set_title( string title);
         void set_description( string description);
         void set_priority( priority tpriority);
@@ -41,20 +41,18 @@ class task{
         void set_repetition_from_string(string tag);
         void set_priority_from_string(string tag);
 
-        int get_identifier()        {return identifier;}
-        string get_title()          {return title;}
-        string get_description()    {return description;}
-        priority get_priority()     {return tpriority;}
-        tdate get_date()            {return date;}
-        ttime get_startingtime()    {return startingtime;}
-        int get_length()            {return length;}
-        repetition get_repetition() {return trepetition;}
-        string get_category()       {return category;}
+        unsigned long get_identifier()  {return identifier;}
+        string get_title()              {return title;}
+        string get_description()        {return description;}
+        priority get_priority()         {return tpriority;}
+        tdate get_date()                {return date;}
+        ttime get_startingtime()        {return startingtime;}
+        int get_length()                {return length;}
+        repetition get_repetition()     {return trepetition;}
+        string get_category()           {return category;}
 
         string get_priority_string();
         string get_repetition_string();
-
-
 };
 
 #endif // TASK_H
